requirements.txt
# ============================================================
# GenAI FMCG Supply Chain Research Assistant — Dependencies
# ============================================================

# --- Core LLM & Agent Framework ---
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0       # Google Gemini (AI Studio) support
langchain-google-vertexai>=2.0.0    # Google Gemini (Vertex AI) support
langchain-community>=0.3.0

# --- Data Validation & Schemas ---
pydantic>=2.5.0
pydantic-settings>=2.1.0

# --- Web Search Tools ---
tavily-python>=0.5.0                 # Tavily search API (recommended)
google-search-results>=2.4.2        # SerpAPI fallback

# --- HTTP & Async ---
httpx>=0.27.0
aiohttp>=3.9.0

# --- Environment & Configuration ---
python-dotenv>=1.0.0

# --- Logging & Observability ---
structlog>=24.1.0

# --- Output & Serialization ---
orjson>=3.9.0                        # Fast JSON serialization

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# --- Development ---
ruff>=0.3.0                          # Linting & formatting
mypy>=1.8.0                          # Type checking

# --- UI & Reporting ---
streamlit>=1.38.0                    # Command Center dashboard
fpdf2>=2.7.0                         # PDF report generation
